pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "netlat"
version = "0.1.0"
description = "Analyze pcap captures for TCP latency, retransmissions, and anomalies"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "typer[all]>=0.9",
    "dpkt>=1.9.8",
    "prometheus-client>=0.20",
    "structlog>=24.0",
    "orjson>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "scapy>=2.5",
    "ruff",
    "mypy",
    "pyyaml>=6.0",
]

[project.scripts]
netlat = "netlat.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
